requirements.txt
torch==1.9.0+cu111
pytorch-lightning==1.4.2
pandas==1.1.4
pyarrow==4.0.0
scikit-learn==0.23.2
hydra-core==1.0.4
GitPython==3.1.11
plotly==4.13.0
statsmodels
onnx
torchvision
fvcore
iopath
pytorch3d@git+https://github.com/facebookresearch/pytorch3d.git@stable
